# contracts/openapi/v1.yaml
openapi: 3.1.0
info:
  title: Policy Orchestrator & Algorithm Execution API
  version: 1.0.0
  description: >-
    Enterprise-grade, vendor-agnostic REST API for repository invariant auditing,
    grounded hybrid GraphRAG, autonomous AI Agent orchestration, and direct execution
    of all 152 Layer 1 algorithms across Search, Observability, Update, Vector, Graph, Filter, and Transform categories.
    Strictly conforms to the RFC API Request & Response Structure Specification (v5.0).
servers:
  - url: http://localhost:8000
    description: Local development orchestrator
paths:
...
  /api/v1/algos/update/patch:
    post:
      summary: Deterministic Atomic Multi-File Patching (Update Algorithm Engine)
      operationId: applyBatchPatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [operations]
              properties:
                operations:
                  type: array
                  items:
                    type: object
                    required: [file_path, find_pattern, replace_text]
                    properties:
                      file_path: { type: string }
                      find_pattern: { type: string }
                      replace_text: { type: string }
                      expected_sha256: { type: string }
                      is_regex: { type: boolean, default: false }
                dry_run:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Batch patch execution and SHA-256 transitions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_AlgoResponse'
  /api/v1/algos/update/diff:
    post:
      summary: Unified GNU/Git Context Diff Engine (Update Algorithm Engine)
      operationId: generateDiff
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [original_content, modified_content]
              properties:
                original_content: { type: string }
                modified_content: { type: string }
                file_path: { type: string, default: "file" }
      responses:
        '200':
          description: Unified diff text and added/deleted counts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_AlgoResponse'
...
  /api/v1/algos/vector-filter/pre-filter:
    post:
      summary: Vector Metadata Pre-Filtering (ALGO-VEC-FLTR-80)
      operationId: vectorFilterPreFilter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vectors, metadata, query, filters]
              properties:
                vectors:
                  type: array
                  items: { type: array, items: { type: number } }
                metadata:
                  type: array
                  items: { type: object }
                query:
                  type: array
                  items: { type: number }
                filters:
                  type: object
                k: { type: integer, default: 5 }
      responses:
        '200':
          description: Pre-filtered search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_AlgoResponse'
  /api/v1/algos/vector-filter/post-filter:
    post:
      summary: Vector Post-Filtering with Oversampling (ALGO-VEC-FLTR-81)
      operationId: vectorFilterPostFilter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vectors, metadata, query, filters]
              properties:
                vectors:
                  type: array
                  items: { type: array, items: { type: number } }
                metadata:
                  type: array
                  items: { type: object }
                query:
                  type: array
                  items: { type: number }
                filters:
                  type: object
                k: { type: integer, default: 5 }
                oversample_factor: { type: number, default: 4.0 }
      responses:
        '200':
          description: Post-filtered search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_AlgoResponse'
  /api/v1/algos/vector-filter/in-graph:
    post:
      summary: ACORN In-Graph Filtering (ALGO-VEC-FLTR-82)
      operationId: vectorFilterInGraph
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vectors, metadata, adjacency, entry_point, query, filters]
              properties:
                vectors:
                  type: array
                  items: { type: array, items: { type: number } }
                metadata:
                  type: array
                  items: { type: object }
                adjacency:
                  type: object
                entry_point: { type: integer, default: 0 }
                query:
                  type: array
                  items: { type: number }
                filters:
                  type: object
                k: { type: integer, default: 5 }
                ef_search: { type: integer, default: 16 }
      responses:
        '200':
          description: In-graph filtered search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_AlgoResponse'
  /api/v1/algos/vector-filter/selectivity-plan:
    post:
      summary: Selectivity-Based Query Planner (ALGO-VEC-FLTR-83)
      operationId: vectorFilterSelectivityPlan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [total_vectors, filters]
              properties:
                total_vectors: { type: integer }
                metadata_sample:
                  type: array
                  items: { type: object }
                filters:
                  type: object
                is_security_filter: { type: boolean, default: false }
      responses:
        '200':
          description: Recommended filtering execution strategy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_AlgoResponse'
  /api/v1/algos/vector-filter/partitioned:
    post:
      summary: Partitioned Multi-Tenant Index Search (ALGO-VEC-FLTR-84)
      operationId: vectorFilterPartitioned
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [partitions, target_partition, query]
              properties:
                partitions:
                  type: object
                target_partition: { type: string }
                query:
                  type: array
                  items: { type: number }
                k: { type: integer, default: 5 }
      responses:
        '200':
          description: Partition search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_AlgoResponse'
...
components:
  schemas:
    ResponseMeta:
      type: object
      required:
        - requestId
        - correlationId
        - timestamp
        - apiVersion
        - status
        - trace_id
      properties:
        requestId:
          type: string
          example: "req-1700000000000-a1b2c3"
        correlationId:
          type: string
          example: "corr-1700000000000-x9y8z7"
        causationId:
          type: string
          nullable: true
        timestamp:
          type: string
          example: "2026-10-05T14:00:00.000Z"
        executionTimeMs:
          type: integer
          example: 12
        apiVersion:
          type: string
          example: "v1"
        status:
          type: string
          enum: [success, error]
        trace_id:
          type: string
          example: "0af7651916cd43dd8448eb211c80319c"
        auditLogId:
          type: string
          nullable: true
        operationId:
          type: string
          nullable: true
        deprecatedFields:
          type: array
          items: { type: string }
          nullable: true
        supportedVersions:
          type: array
          items: { type: string }
        staleness:
          type: object
          nullable: true
        pagination:
          type: object
          nullable: true
        links:
          type: object
          nullable: true
    ErrorContainer:
      type: object
      required:
        - code
        - message
        - retryable
      properties:
        code:
          type: string
          example: "VALIDATION_FAILED"
        message:
          type: string
          example: "One or more payload validation checks failed."
        retryable:
          type: boolean
          default: false
        details:
          type: array
          items:
            type: object
    ErrorItem:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: "VALIDATION_FAILED"
        message:
          type: string
          example: "One or more payload validation checks failed."
        details:
          type: object
...
    Envelope_AlgoResponse:
      type: object
      required: [success, statusCode, meta, data]
      properties:
        success: { type: boolean, example: true }
        statusCode: { type: integer, example: 200 }
        data: { type: object }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
        error: { $ref: '#/components/schemas/ErrorContainer' }
        errors: { type: array, items: { $ref: '#/components/schemas/ErrorItem' }, nullable: true }
...